Tuple
When you want to return multiple values from a method without using ref or out parameters.
1. To pass multiple values to a method through a single parameter.
To hold a database record or some values temporarily without creating a separate class.
2.Tuple Limitations:
The Tuple is a reference type and not a value type. 
The Tuple is limited to include eight elements. You need to use nested tuples if you need to store more elements.